Move the delete and archive button to PhotoCard component
