+++ src/components/PhotoCard.js
@@ -1,46 +1,68 @@
 import PropTypes from 'prop-types';
-import { NavLink } from 'react-router-dom';
-import DeletePhoto from './DeletePhoto';
-import ArchivePhoto from './ArchivePhoto';
+import { NavLink, useNavigate } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 import LikePhoto from './LikePhoto';
+import { archivePhoto, deletePhoto } from '../redux/posts/photoSlice';
 
 const PhotoCard = ({
   id, name, imageUrl, caption, likes, liked, control, navigator, archive,
-}) => (
-  <div>
-    <p>
-      {name}
-    </p>
-    {navigator ? (
-      <NavLink to={`photos/${id}`}>
-        <img src={imageUrl} alt="i-share" />
-      </NavLink>
-    )
-      : <img src={imageUrl} alt="i-share" />}
-    <p>
-      {caption}
-    </p>
-    <span>
-      {likes}
-    </span>
-    <LikePhoto
-      id={id}
-      liked={liked}
-    />
-    {control && (
+}) => {
+  const dispatch = useDispatch();
+  const handleArchive = async () => {
+    try {
+      await dispatch(archivePhoto(id));
+    } catch (error) {
+      console.error(error);
+    }
+  };
+  const navigate = useNavigate();
+  const handleDelete = async () => {
+    try {
+      const { payload } = await dispatch(deletePhoto(id));
+
+      if (payload && payload.deleted) {
+        navigate('/');
+      }
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  return (
+    <div>
+      <p>
+        {name}
+      </p>
+      {navigator ? (
+        <NavLink to={`photos/${id}`}>
+          <img src={imageUrl} alt="i-share" />
+        </NavLink>
+      )
+        : <img src={imageUrl} alt="i-share" />}
+      <p>
+        {caption}
+      </p>
+      <span>
+        {likes}
+      </span>
+      <LikePhoto
+        id={id}
+        liked={liked}
+      />
+      {control && (
       <>
-        <DeletePhoto
-          id={id}
-        />
-        <ArchivePhoto
-          id={id}
-          archive={archive}
-        />
+        <button type="button" onClick={handleDelete}>
+          Delete
+        </button>
+        <button type="button" onClick={handleArchive}>
+          {archive ? 'Unarchive' : 'Archive'}
+        </button>
       </>
 
-    )}
-  </div>
-);
+      )}
+    </div>
+  );
+};
 export default PhotoCard;
 
 PhotoCard.propTypes = {
